+++ src/native/java_class_ids.c
@@ -1912,6 +1912,22 @@ static void s_cache_mqtt5_publish_events_properties(JNIEnv *env) {
     AWS_FATAL_ASSERT(mqtt5_publish_events_properties.publish_events_publish_received_id);
 }
 
+struct java_aws_mqtt5_listener_publish_events mqtt5_listener_publish_events_properties;
+
+static void s_cache_mqtt5_listener_publish_events_properties(JNIEnv *env) {
+    jclass cls = (*env)->FindClass(env, "software/amazon/awssdk/crt/mqtt5/Mqtt5ListenerOptions$ListenerPublishEvents");
+    AWS_FATAL_ASSERT(cls);
+    mqtt5_listener_publish_events_properties.listener_publish_events_class = (*env)->NewGlobalRef(env, cls);
+    AWS_FATAL_ASSERT(mqtt5_listener_publish_events_properties.listener_publish_events_class);
+    // Functions
+    mqtt5_listener_publish_events_properties.listener_publish_events_publish_received_id = (*env)->GetMethodID(
+        env,
+        mqtt5_listener_publish_events_properties.listener_publish_events_class,
+        "onMessageReceived",
+        "(Lsoftware/amazon/awssdk/crt/mqtt5/Mqtt5Client;Lsoftware/amazon/awssdk/crt/mqtt5/PublishReturn;)Z");
+    AWS_FATAL_ASSERT(mqtt5_listener_publish_events_properties.listener_publish_events_publish_received_id);
+}
+
 struct java_aws_mqtt5_lifecycle_events mqtt5_lifecycle_events_properties;
 
 static void s_cache_mqtt5_lifecycle_events_properties(JNIEnv *env) {
@@ -2083,7 +2099,7 @@ static void s_cache_mqtt5_listener_options(JNIEnv *env) {
         env,
         mqtt5_listener_options_properties.listener_options_class,
         "lifecycleEvents",
-        "Lsoftware/amazon/awssdk/crt/mqtt5/Mqtt5ListenerOptions$LifecycleEvents;");
+        "Lsoftware/amazon/awssdk/crt/mqtt5/Mqtt5ClientOptions$LifecycleEvents;");
     AWS_FATAL_ASSERT(mqtt5_listener_options_properties.lifecycle_events_field_id);
 }
 
@@ -2234,6 +2250,7 @@ void cache_java_class_ids(JNIEnv *env) {
     s_cache_mqtt5_unsuback_reason_code(env);
     s_cache_mqtt5_user_property(env);
     s_cache_mqtt5_publish_events_properties(env);
+    s_cache_mqtt5_listener_publish_events_properties(env);
     s_cache_mqtt5_lifecycle_events_properties(env);
     s_cache_mqtt5_puback_result(env);
     s_cache_mqtt5_publish_return(env);
